move mock components to their own module and separate reader/writer
